+++ app/src/main/java/illyan/jay/data/disk/model/RoomSession.kt
@@ -5,9 +5,9 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "session")
 data class RoomSession(
-    @PrimaryKey(autoGenerate = true)
-    val id: Int = 0,
-    val startTime: Long,
-    var endTime: Long? = null,
-    var distance: Double = 0.0
+	@PrimaryKey(autoGenerate = true)
+	val id: Long = 0,
+	val startTime: Long,
+	var endTime: Long? = null,
+	var distance: Double = 0.0
 )
