+++ app/src/main/java/illyan/jay/domain/model/DomainAcceleration.kt
@@ -4,11 +4,11 @@ import java.util.*
 
 
 data class DomainAcceleration(
-    val id: Int = -1,
-    val sessionId: Int,
-    val time: Date,
-    val accuracy: Int, // enum
-    val x: Float,
-    val y: Float,
-    val z: Float
+	val id: Long = -1,
+	val sessionId: Long,
+	val time: Date,
+	val accuracy: Int, // enum
+	val x: Float,
+	val y: Float,
+	val z: Float
 )
