+++ app/src/main/java/illyan/jay/domain/model/DomainRotation.kt
@@ -3,11 +3,11 @@ package illyan.jay.domain.model
 import java.util.*
 
 data class DomainRotation(
-    val id: Int = -1,
-    val sessionId: Int,
-    val time: Date,
-    val accuracy: Int, // enum
-    val x: Float,
-    val y: Float,
-    val z: Float
+	val id: Long = -1,
+	val sessionId: Long,
+	val time: Date,
+	val accuracy: Int, // enum
+	val x: Float,
+	val y: Float,
+	val z: Float
 )
